+++ apps/api/src/admin-auth/services/password.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { hashPassword, verifyPassword, verifyPasswordTimingSafe } from '../utils/password.util';
+import { hashPassword, verifyPassword, verifyPasswordTimingSafe } from '../../common/crypto/password.util';
 
 @Injectable()
 export class PasswordService {